requirements.txt
fastapi[standard]>=0.115.12
uvicorn>=0.27.1
python-multipart>=0.0.6
streamlit>=1.32.0
faiss-cpu>=1.7.4
sentence-transformers>=2.5.1
langchain>=0.1.9
langchain-community>=0.0.24
python-dotenv>=1.0.1
pydantic>=2.6.3
groq>=0.4.2
requests>=2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
tiktoken>=0.5.1 